Trees aka potential parking lot spots #11
Implemented the MakeTree class
Changed the tree concept: it works on and arraylist of objects that implement the Node interface
Made RawDictionary implement Node interface (i.e. it's a leaf)
Made ReadRawDictionary return and arraylist of objects (for the MakeTree)
Removed the Dictionary class because MakeTree already does what it was going to do
Added Node interface
Added BlankNode class (a.k.a. "branch" or node without a symbol)